requirements.txt
recordtype==1.3
requests==2.19.1
py-postgresql==1.2.1
